refactor: exception 패키지 재구성 및 네이밍 변경
