+++ src/main/java/woowacourse/shoppingcart/application/CartService.java
@@ -11,8 +11,8 @@ import woowacourse.shoppingcart.dao.CustomerDao;
 import woowacourse.shoppingcart.dao.ProductDao;
 import woowacourse.shoppingcart.domain.Cart;
 import woowacourse.shoppingcart.domain.Product;
-import woowacourse.shoppingcart.exception.InvalidProductException;
-import woowacourse.shoppingcart.exception.NotInCustomerCartItemException;
+import woowacourse.shoppingcart.exception.domain.InvalidProductException;
+import woowacourse.shoppingcart.exception.domain.NotInCustomerCartItemException;
 
 @Service
 @Transactional(rollbackFor = Exception.class)
