+++ src/main/java/woowacourse/shoppingcart/application/CustomerService.java
@@ -10,8 +10,8 @@ import woowacourse.shoppingcart.dto.CustomerRequest;
 import woowacourse.shoppingcart.dto.CustomerResponse;
 import woowacourse.shoppingcart.dto.FindCustomerRequest;
 import woowacourse.shoppingcart.dto.UpdateCustomerRequest;
-import woowacourse.shoppingcart.exception.DuplicateCustomerException;
-import woowacourse.shoppingcart.exception.InvalidCustomerException;
+import woowacourse.shoppingcart.exception.domain.DuplicateCustomerException;
+import woowacourse.shoppingcart.exception.domain.CustomerNotFoundException;
 
 @Service
 public class CustomerService {
@@ -39,13 +39,13 @@ public class CustomerService {
 
     public CustomerResponse findCustomer(FindCustomerRequest findCustomerRequest) {
         Customer customer = customerDao.findByName(findCustomerRequest.getName())
-            .orElseThrow(InvalidCustomerException::new);
+            .orElseThrow(CustomerNotFoundException::new);
         return CustomerResponse.from(customer);
     }
 
     public void updateCustomer(FindCustomerRequest findCustomerRequest, UpdateCustomerRequest updateCustomerRequest) {
         Customer customer = customerDao.findByName(findCustomerRequest.getName())
-            .orElseThrow(InvalidCustomerException::new);
+            .orElseThrow(CustomerNotFoundException::new);
         Customer updatedCustomer = customer.update(updateCustomerRequest.getAddress(),
             updateCustomerRequest.getPhoneNumber());
         customerDao.update(updatedCustomer);
