+++ src/main/java/woowacourse/shoppingcart/application/OrderService.java
@@ -16,7 +16,7 @@ import woowacourse.shoppingcart.domain.OrderDetail;
 import woowacourse.shoppingcart.domain.Orders;
 import woowacourse.shoppingcart.domain.Product;
 import woowacourse.shoppingcart.dto.OrderRequest;
-import woowacourse.shoppingcart.exception.InvalidOrderException;
+import woowacourse.shoppingcart.exception.domain.InvalidOrderException;
 
 @Service
 @Transactional(rollbackFor = Exception.class)
