+++ src/main/java/woowacourse/shoppingcart/dao/ProductDao.java
@@ -10,7 +10,7 @@ import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.stereotype.Repository;
 
 import woowacourse.shoppingcart.domain.Product;
-import woowacourse.shoppingcart.exception.InvalidProductException;
+import woowacourse.shoppingcart.exception.domain.InvalidProductException;
 
 @Repository
 public class ProductDao {
