+++ src/main/java/woowacourse/shoppingcart/exception/domain/CartItemNotFoundException.java
@@ -1,0 +1,12 @@
+package woowacourse.shoppingcart.exception.domain;
+
+public class CartItemNotFoundException extends NotFoundException {
+
+    public CartItemNotFoundException() {
+        this("유효하지 않은 장바구니입니다.");
+    }
+
+    public CartItemNotFoundException(final String msg) {
+        super(msg);
+    }
+}
