+++ src/main/java/woowacourse/shoppingcart/exception/domain/CustomerNotFoundException.java
@@ -1,0 +1,12 @@
+package woowacourse.shoppingcart.exception.domain;
+
+public class CustomerNotFoundException extends NotFoundException {
+
+    public CustomerNotFoundException() {
+        this("존재하지 않는 유저입니다.");
+    }
+
+    public CustomerNotFoundException(final String msg) {
+        super(msg);
+    }
+}
