+++ src/main/java/woowacourse/shoppingcart/exception/domain/DuplicateCustomerException.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.exception;
+package woowacourse.shoppingcart.exception.domain;
 
 public class DuplicateCustomerException extends DuplicateException {
