+++ src/test/java/woowacourse/shoppingcart/domain/AddressTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import woowacourse.shoppingcart.domain.customer.Address;
-import woowacourse.shoppingcart.exception.InvalidLengthException;
+import woowacourse.shoppingcart.exception.attribute.InvalidLengthException;
 
 class AddressTest {
